removes pynvml, which makes ROCm possible
pytorch has built in functions to get the devices and used memory. If we
use them instead of pynvml, it runs fine on my AMD card using ROCm. It
needs the module amdsmi, but the error message is helpful enough, so
that people without nvidia gpus can install it themselves. Maybe this
module can get added somewhere as a dependency.